[consensus] ensure the reset waits for all ongoing tasks finish
It's a race condition when PersistingPhase has an ongoing request to commit while reset is acknowledged and switch to state sync.
And state sync will try to commit to db as well and result in corrupted db.

We'll make sure db is thread-safe in a separate commit, but this ensures we avoid this specefic race condition.

Closes: #1477