[consensus] Create ProposerAndVoterHeuristic
Heuristic that looks at successful and failed proposals, as well as voting history,
to define node reputation, used for leader selection.

We want to optimize leader selection to primarily maximize network's throughput,
but we also, in combinatoin with staking rewards logic, need to be reasonably fair.

Logic is:
 * if proposer round failure rate within the proposer window is strictly above threshold, use failed_weight (default 1).
 * otherwise, if node had no proposal rounds and no successful votes, use inactive_weight (default 10).
 * otherwise, use the default active_weight (default 100).

We primarily want to avoid failed rounds, as they have a largest negative effect on the network.
So if we see a node having failures to propose, when it was the leader, we want to avoid that node.
We add a threshold (instead of penalizing on a single failure), so that transient issues in the network,
or malicious behaviour of the next leader is avoided. In general, we expect there to be
proposer_window_size/num_validators opportunities for a node to be a leader, so a single failure, or a
subset of following leaders being malicious will not be enough to exclude a node.
On the other hand, single failure, without any successes before will exclude the note.
Threshold probably makes the most sense to be between:
 * 10% (aggressive exclusion with 1 failure in 10 proposals being enough for exclusion)
 * and 33% (much less aggressive exclusion, with 1 failure for every 2 successes, should still reduce failed
   rounds by at least 66%, and is enough to avoid byzantine attacks as well as the rest of the protocol)

Closes: #1395